chore: update rustls (#192)
Update `rustls` to use `async-rustls` which makes it compatible with async.  
Remove `rustls` and `webpki` from dependency since `async-rustls` re-export.
Bump version to 0.4.4

+++ src/zero_copy.rs
@@ -125,7 +125,7 @@ impl ZeroCopy {
                 use nix::errno::Errno;
 
                 let mut total_transferred = 0;
-                let mut current_offset = offset as u64;
+                let mut current_offset = offset;
 
                 loop {
                     let to_be_transfer = (size - total_transferred) as i64;
